[spa] fix breakages from upgrade to angular 9
Chiefly, to prevent double key-presses by only having one keydown listener
