+++ pupil-spa/src/app/practice-question/practice-question.component.spec.ts
@@ -1,16 +1,18 @@
 import { async, fakeAsync, tick, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { PracticeQuestionComponent } from './practice-question.component';
+import { AnswerService } from '../services/answer/answer.service';
 import { AuditService } from '../services/audit/audit.service';
 import { AuditServiceMock } from '../services/audit/audit.service.mock';
+import { PracticeQuestionComponent } from './practice-question.component';
+import { QuestionService } from '../services/question/question.service';
+import { QuestionServiceMock } from '../services/question/question.service.mock';
+import { RegisterInputService } from '../services/register-input/registerInput.service';
+import { RegisterInputServiceMock } from '../services/register-input/register-input-service.mock';
+import { SoundComponentMock } from '../sound/sound-component-mock';
 import { SpeechService } from '../services/speech/speech.service';
 import { SpeechServiceMock } from '../services/speech/speech.service.mock';
 import { StorageService } from '../services/storage/storage.service';
-import { QuestionService } from '../services/question/question.service';
-import { QuestionServiceMock } from '../services/question/question.service.mock';
 import { WindowRefService } from '../services/window-ref/window-ref.service';
-import { SoundComponentMock } from '../sound/sound-component-mock';
-import {AnswerService} from '../services/answer/answer.service';
 
 describe('PractiseQuestionComponent', () => {
   let component: PracticeQuestionComponent;
@@ -26,6 +28,7 @@ describe('PractiseQuestionComponent', () => {
         { provide: AuditService, useClass: AuditServiceMock },
         { provide: SpeechService, useValue: mockSpeechService },
         { provide: QuestionService, useClass: QuestionServiceMock },
+        { provide: RegisterInputService, useClass: RegisterInputServiceMock },
         StorageService,
         WindowRefService,
         AnswerService
