+++ pupil-spa/src/app/spoken-practice-question/spoken-practice-question.component.ts
@@ -2,13 +2,14 @@ import { Component, NgZone, OnInit, AfterViewInit, OnDestroy, Input } from '@ang
 import { Subscription } from 'rxjs';
 
 import { AnswerService } from '../services/answer/answer.service';
-import { PracticeQuestionComponent } from '../practice-question/practice-question.component';
 import { AuditService } from '../services/audit/audit.service';
-import { WindowRefService } from '../services/window-ref/window-ref.service';
-import { SpeechService } from '../services/speech/speech.service';
-import { StorageService } from '../services/storage/storage.service';
+import { PracticeQuestionComponent } from '../practice-question/practice-question.component';
 import { QuestionRendered } from '../services/audit/auditEntry';
 import { QuestionService } from '../services/question/question.service';
+import { RegisterInputService } from '../services/register-input/registerInput.service';
+import { SpeechService } from '../services/speech/speech.service';
+import { StorageService } from '../services/storage/storage.service';
+import { WindowRefService } from '../services/window-ref/window-ref.service';
 
 @Component({
   selector: 'app-spoken-practice-question',
@@ -24,8 +25,9 @@ export class SpokenPracticeQuestionComponent extends PracticeQuestionComponent i
               protected zone: NgZone,
               protected storageService: StorageService,
               protected questionService: QuestionService,
-              protected answerService: AnswerService) {
-    super(auditService, windowRefService, questionService, storageService, speechService, answerService);
+              protected answerService: AnswerService,
+              protected registerInputService: RegisterInputService) {
+    super(auditService, windowRefService, questionService, storageService, speechService, answerService, registerInputService);
     this.isWarmUpQuestion = true;
   }
